refactor: frontend project structure
