+++ frontend/src/lib/components/header.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { page } from '$app/state';
-  import ChartBarsIcon from '$lib/components/icons/chart_bars.svelte';
+  import ChartBarsIcon from '$lib/components/icons/chart-bars.svelte';
+  import QuibbleTextLogo from '$lib/components/icons/logos/quibble-text.svelte';
   import QuibbleLogo from '$lib/components/icons/logos/quibble.svelte';
-  import QuibbleTextLogo from '$lib/components/icons/logos/quibble_text.svelte';
   import NotificationIcon from '$lib/components/icons/notification.svelte';
   import Avatar from '$lib/components/ui/avatar.svelte';
   import { createAuthStore } from '$lib/stores/auth.svelte';
