+++ frontend/src/lib/components/icons/chart-bars.svelte
