+++ frontend/src/lib/components/icons/logos/quibble-text.svelte
