--- frontend/src/lib/components/modals/index.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-  import ModalAuth from './modal_auth/index.svelte';
-  import ModalCreateCommunity from './modal_create_community/index.svelte';
-</script>
-
-<ModalAuth />
-<ModalCreateCommunity />
