--- frontend/src/lib/components/modals/modal_auth/forms/index.ts
@@ -1,7 +1,0 @@
-const forms = {
-  join: import('./join.svelte'),
-  profile_select: import('./profile_select.svelte'),
-  profile_create: import('./profile_create.svelte')
-};
-
-export default forms;
