+++ frontend/src/lib/components/post.svelte
@@ -2,7 +2,7 @@
   import { page } from '$app/state';
   import type { components } from '$lib/clients/v1';
   import Avatar from '$lib/components/ui/avatar.svelte';
-  import BackdropImage from '$lib/components/ui/backdrop_image.svelte';
+  import BackdropImage from '$lib/components/ui/backdrop-image.svelte';
   import Zoom from '$lib/components/ui/zoom.svelte';
   import { cn } from '$lib/functions/classnames';
   import { FormatDate } from '$lib/functions/date';
