+++ frontend/src/lib/components/ui/backdrop-image.svelte
