+++ frontend/src/lib/components/vectors/quibble-404-2.svelte
