+++ frontend/src/lib/features/comment/index.ts
@@ -1,0 +1,1 @@
+export { default as CommentBlock } from './comment-block.svelte';
