+++ frontend/src/lib/functions/is-valid.ts
