+++ frontend/src/lib/modals/_components/base-modal.svelte
