+++ frontend/src/lib/modals/index.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import ModalAuth from './modal-auth/index.svelte';
+  import ModalCreateCommunity from './modal-create-community/index.svelte';
+</script>
+
+<ModalAuth />
+<ModalCreateCommunity />
