+++ frontend/src/lib/modals/modal-auth/forms/index.ts
@@ -1,0 +1,7 @@
+const forms = {
+  join: import('./join.svelte'),
+  profile_select: import('./profile-select.svelte'),
+  profile_create: import('./profile-create.svelte')
+};
+
+export default forms;
