+++ frontend/src/lib/modals/modal-auth/forms/profile-create.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { enhance } from '$app/forms';
+  import QuibbleTextLogo from '$lib/components/icons/logos/quibble-text.svelte';
   import QuibbleLogo from '$lib/components/icons/logos/quibble.svelte';
-  import QuibbleTextLogo from '$lib/components/icons/logos/quibble_text.svelte';
-  import ZodErrors from '$lib/components/shared/zod-errors.svelte';
+  import ZodErrors from '$lib/components/zod-errors.svelte';
   import { cn } from '$lib/functions/classnames';
   import type { FormProps } from '../../types';
   import forms from '../forms';
