+++ frontend/src/lib/modals/modal-create-community/index.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { cn } from '$lib/functions/classnames';
   import { createModalsStore } from '$lib/stores/modals.svelte';
-  import BaseModal from '../_components/base_modal.svelte';
+  import BaseModal from '../_components/base-modal.svelte';
   import { create_form_history } from '../_utils/history.svelte';
   import type { FormsState, FormSubmitData, Forms } from '../types';
   import forms from './forms';
