+++ frontend/src/routes/(app)/q/[name]/posts/[id]/[slug]/+page.server.ts
@@ -1,4 +1,4 @@
-import client from '$lib/clients/v1/client';
+import client from '$lib/clients/client';
 import { CommentTreeBuilder } from '$lib/functions/comment';
 import type { PageServerLoad } from './$types';
 import { error as raise_error, redirect } from '@sveltejs/kit';
