+++ frontend/src/routes/(app)/q/[name]/posts/[id]/[slug]/+page.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import { browser } from '$app/environment';
-  import CommentBlock from '$lib/components/comment_block.svelte';
   import NewIcon from '$lib/components/icons/new.svelte';
   import RocketIcon from '$lib/components/icons/rocket.svelte';
   import TopIcon from '$lib/components/icons/top.svelte';
   import Avatar from '$lib/components/ui/avatar.svelte';
-  import BackdropImage from '$lib/components/ui/backdrop_image.svelte';
+  import BackdropImage from '$lib/components/ui/backdrop-image.svelte';
   import Zoom from '$lib/components/ui/zoom.svelte';
+  import { CommentBlock } from '$lib/features/comment';
   import { cn } from '$lib/functions/classnames';
   import { FormatDate } from '$lib/functions/date';
   import { is_valid } from '$lib/functions/is_valid';
